Stable Market
Sometimes, the stability of a stock market is at least as important as rapid price changes.

Let a stable block be a maximal consecutive block of days with same stock prices.

Moreover, let a stable block of order K be a stable block of length at least K.

For example, if stock prices for 7 consecutive days are: 20, 10, 10, 7, 7, 7, 10, then there are 4 stable blocks there: [20], [10, 10], [7, 7, 7] and [10]. Moreover, there are:
0 stable blocks of order 4 or higher
1 stable block of order 3: [7, 7, 7]
2 stable block of order 2: [10, 10], [7, 7, 7]
4 stable block of order 1: [20], [10, 10], [7, 7, 7] and [10]
Given historical prices of a stock across N days, the goal is to answer Q customers' questions.

The i-th question contains three integers Li, Ri and Ki. It asks for the number of stable blocks of order Ki, if only the stock prices of the days in the range Li, Ri are considered.